ReadMe
The training of the models decision tree, random forest, svm, k-nearest neighbor (models originally from https://www.kaggle.com/mayu0116/hyper-parameters-tuning-of-dtree-rf-svm-knn) and the ensemble of the four models with the soft voting classifier is done in the Jupyer notebook 'Breast Cancer Code.ipybn'. In the same notebook the ROC plots and the accuracy table mentioned in the report are created. (The notebook also generates plots for the training data, which are not placed in the report.)

The data used in the notebook is stored as 'dataset_180_408_data.csv', similar to the kaggle competition: https://www.kaggle.com/uciml/breast-cancer-wisconsin-data.